cpld: Bugfix: burst writes sometimes cause FSM false command execution
Burst writes (writes of longer than 1 word) occasionally caused invalid commands to be executed.
This often manifested as LED state changes / mode changes. (this is more obvious if the LEDs are reconfigured to show `exfiltration_enabled`, `mode[1]` and `mode[0]`).

Add an extra delay to account for the slight delay while the FT240X makes RXF inactive after the last byte is read.